LayoutTests/http/tests/security/contentSecurityPolicy/inline-script-blocked-goofy-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'options'.

CONSOLE MESSAGE: Refused to execute inline script because of Content-Security-Policy.

CONSOLE MESSAGE: Refused to execute inline event handler because of Content-Security-Policy.

This test passes if it doesn't alert fail.
